+++ completion/__init__.py
@@ -1,0 +1,1 @@
+# See https://github.com/python/cpython/issues/89838 for why this file is needed